requirements.txt
streamlit
langchain
langchain-google-genai>=1.0.3
google-generativeai>=0.5.2
python-dotenv
pymupdf
python-docx
pandas
sqlalchemy
passlib
bcrypt
plotly
